+++ src/LaunchDarkly.CommonSdk/Internal/IHttpRequestConfiguration.cs
@@ -1,0 +1,13 @@
+﻿using System.Net.Http;
+
+namespace LaunchDarkly.Common
+{
+    /// <summary>
+    /// Used internally to describe all configuration properties that affect HTTP requests in general.
+    /// </summary>
+    internal interface IHttpRequestConfiguration
+    {
+        string HttpAuthorizationKey { get; }
+        HttpMessageHandler HttpMessageHandler { get; }
+    }
+}