+++ test/LaunchDarkly.CommonSdk.Tests/DefaultEventProcessorTest.cs
@@ -38,7 +38,7 @@ namespace LaunchDarkly.Common.Tests
             }
         }
 
-        private IEventProcessor MakeProcessor(IBaseConfiguration config)
+        private IEventProcessor MakeProcessor(SimpleConfiguration config)
         {
             return new DefaultEventProcessor(config, new TestUserDeduplicator(),
                 Util.MakeHttpClient(config, SimpleClientEnvironment.Instance), EventsUriPath);
@@ -541,7 +541,7 @@ namespace LaunchDarkly.Common.Tests
         {
             if (userJson != null)
             {
-                Assert.Equal(userJson, o["user"]);
+                TestUtil.AssertJsonEquals(userJson, o["user"]);
                 Assert.Null(o["userKey"]);
             }
             else